refactor: split ManageShifts into focused sub-components (Phase 4)
ManageShifts was a 662-line page mixing role-scoped data fetching,
table rendering, a 180-line create/edit dialog, and a delete
confirmation flow. This PR keeps the page as a thin orchestrator
(110 lines) and moves the rest into 5 focused files. Behavior is
unchanged — the existing 103 vitest tests all pass.

Settings-style pattern: the page coordinates dialog open/close
because the shifts table triggers edit/delete/invite actions, but
the dialogs themselves own their internal form/loading state.

New files:
  src/hooks/useShiftsList.ts                    107  Role-scoped fetch
                                                     for shifts +
                                                     departments
  src/lib/shift-validation.ts                   113  Pure: canEditNote,
                                                     computeDurationInfo,
                                                     validateShiftForm
                                                     (discriminated-union
                                                     return), constants
  src/components/shifts/ShiftsTable.tsx          97  Presentational table
                                                     with per-row Invite/
                                                     Edit/Delete buttons
  src/components/shifts/ShiftFormDialog.tsx     322  Create/edit dialog —
                                                     panels-own-state form
                                                     + supabase save
  src/components/shifts/DeleteShiftDialog.tsx    53  Confirm AlertDialog

Page: 662 → 110 (-83%)
Total: 662 → 802 (+140 net; ~550 lines of logic moved out).

Type strategy:
- Shift / Department / ShiftForm / DurationInfo interfaces relocated
  to the new files.
- One new boundary cast in new code (useShiftsList:60) — `as any[]`
  on the embedded department_coordinators → departments select.
  Same documented pattern as PRs #125 / #127.
- Validation logic uses a discriminated-union return so toast strings
  stay in the page's UI layer (long-shift confirm via window.confirm
  remains in the dialog component).

Follow-up issue filed: #128 (ManageShifts spinner hangs if role
resolves to null — pre-existing edge case prevented by the route
guard, not fixed in this behavior-preservation refactor).

Verification: tsc --build clean, eslint --max-warnings=0 clean,
103/103 vitest passes.

Third of five Phase 4 god-component refactors. AdminUsers.tsx
(~650 lines) is next.
